requirements.txt
streamlit
streamlit_code_editor
streamlit-js-eval
openai
jupyter_client
ipykernel
jupyter_client
torch
torchvision
torchaudio